+++ DiscCheckEmuAPIHook/Hook/GetVolumeInformationA.h
@@ -75,13 +75,16 @@ BOOL WINAPI HookedGetVolumeInformationA(LPCSTR lpRootPathName,
         }
     }
 
-    SPDLOG_INFO("<--- GetVolumeInformationA({0:s}, {1:s}, 0x{2:x}, {3:s}, {4:s}, {5:s}, {6:s}, 0x{7:x})",
-        lpRootPathName, "lpVolumeNameBuffer", nVolumeNameSize, "lpVolumeSerialNumber",
-        "lpMaximumComponentLength", "lpFileSystemFlags", "lpFileSystemNameBuffer", nFileSystemNameSize);
-
-    return OGGetVolumeInformationA(lpRootPathName,
+    bool ogResult = OGGetVolumeInformationA(lpRootPathName,
         lpVolumeNameBuffer, nVolumeNameSize,
         lpVolumeSerialNumber, lpMaximumComponentLength,
         lpFileSystemFlags, lpFileSystemNameBuffer,
         nFileSystemNameSize);
+
+    SPDLOG_INFO("<--- GetVolumeInformationA({0:s}, {1:s}, 0x{2:x}, 0x{3:x}, 0x{4:x}, 0x{5:x}, {6:s}, 0x{7:x}) | Returning: {8}",
+        lpRootPathName, lpVolumeNameBuffer, nVolumeNameSize, (lpVolumeSerialNumber == nullptr ? 0 : *lpVolumeSerialNumber),
+        (lpMaximumComponentLength == nullptr ? 0 : *lpMaximumComponentLength), (lpFileSystemFlags == nullptr ? 0 : *lpFileSystemFlags),
+        (lpFileSystemNameBuffer == nullptr ? "NULL" : lpFileSystemNameBuffer), nFileSystemNameSize, ogResult);
+
+    return ogResult;
 }